Final version with video and report
