--- .gitignore
@@ -1,9 +1,0 @@
-bin/
-.vscode/
-*.txt
-.idea/
-example/
-# Clion
-CMakeLists.txt
-cmake-build-debug/
-renan_cpm.sh
