+++ Grupo 6/.gitignore
@@ -1,0 +1,8 @@
+bin/
+.vscode/
+.idea/
+example/
+# Clion
+CMakeLists.txt
+cmake-build-debug/
+renan_cpm.sh
