+++ Grupo 6/run_mutate.sh
@@ -1,0 +1,1 @@
+./bin/mutate_program --inputFileName ./example/chry.txt --probability 20 --outputFileName ./example/mutated_chry.txt --bufferSize 1024
