+++ Grupo 6/src/include/utils/Helpers.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include <cmath>
+
+namespace UTILS {
+
+    class Helpers {
+    public:
+        static double calculateEntropy(double frequency);
+    };
+
+} // UTILS
+
