--- README.md
@@ -1,47 +1,0 @@
-# TAI
-
-Trabalho de Teoria Algorítmica da informação
-
-## Introduction and Project Idea
-
-### Copy Model
-
-## Overview
-
-## Features
-
-## How to use
-
-This section contains steps that one can use to get their own copy of our programs.
-
-- Start by forking this repository using the fork button at the top right corner of the page to get your own copy of the repository. This is to ensure that you can actively experiment with the code of the project.
-
-- Once the repository has been forked, you can clone the repository to your local machine using the following command:
-
-	```bash
-	git clone <paste URL of the forked repo here>
-	```
-
-- Lastly, with the repository cloned, you can navigate to the directory of the repository and run the following command to compile the code:
-
-    ```bash
-    make all
-    ```
-
-    Remember to get 'chry.txt'
-
-This will create the executables by compiling and linking the project. You can now test and run the appropriate shell script to test the programs.
-
-- **Note:** The use of a Linux based system is recommended.
-
-- **Note:** To run the programs on a windows system, you can use the windows subsystem for linux (WSL2).
-
-## Working of the Project
-
-### Basic idea of the working
-
-### In detail working
-
-## Future Work
-
-### Proudly brought to you by our group ❤️
