--- run_cpm.sh
@@ -1,1 +1,0 @@
-./bin/cpm_model --alpha 1 --k 15 --threshold 14 --inputFileName chry.txt --bufSize 1024
