--- run_mutate.sh
@@ -1,1 +1,0 @@
-./bin/mutate_program --inputFileName chry.txt --probability 20 --outputFileName mutated_chry.txt --bufferSize 1024
