--- src/include/utils/Helpers.cpp
@@ -1,7 +1,0 @@
-#include "Helpers.hpp"
-
-namespace UTILS {
-    double Helpers::calculateEntropy(double frequency) {
-        return -frequency*log2(frequency);
-    }
-} // UTILS
