--- src/include/utils/Helpers.hpp
@@ -1,12 +1,0 @@
-#pragma once
-#include <cmath>
-
-namespace UTILS {
-
-    class Helpers {
-    public:
-        static double calculateEntropy(double frequency);
-    };
-
-} // UTILS
-
